- Plugin: Fix for Warn $testStatus value showing as error in UI - Plugin: Fix DupArchive engine to add failure when fopen returns null - Plugin: Move procedure privileges scan result to Database section - Plugin: Write in log last sql file lines if mysqldump fail - Plugin: Include wp-content/bps-backup folder into the global folder filter list - Installer: Add support for hashed installer name - Installer: Console error while downloading a installer log - Installer: jQuery loading issues on some systems, links are now external
git-svn-id: https://plugins.svn.wordpress.org/duplicator/trunk@2298564 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ views/tools/diagnostics/information.php
@@ -76,7 +76,12 @@ if ($section == "info" || $section == '') {
 						$file_path = '';
 						if (stripos($filename, '[hash]') !== false) {
 							$glob_files = glob($path);
+                            
 							if (!empty($glob_files)) {
+                                if(count($glob_files) > 10) {                                
+                                    throw new Exception('Trying to delete too many files. Please contact Duplicator support.');
+                                }
+                                
 								foreach ($glob_files as $glob_file) {
 									$file_path = $glob_file;
 									DUP_IO::deleteFile($file_path);
@@ -95,7 +100,7 @@ if ($section == "info" || $section == '') {
 								DUP_IO::deleteTree($path);
 								$removed_files = true;
 							}
-						}
+						}                            
 
 						if (!empty($file_path)) {
                             if (file_exists($file_path)) {
